refactors dataset generation
